Fixed bug in DB where the data would get corrupted due to float64 being saved and not float32

+++ smocs/db/mysql_api_v0.py
@@ -209,7 +209,7 @@ class DBManager:
                     result[key] = float(result[key])
                 if type(result[key]) is bytes:
                     try:
-                        result[key] = np.frombuffer(result[key], dtype=np.float64)
+                        result[key] = np.frombuffer(result[key], dtype=np.float32)
                     except ValueError:
                         try:
                             result[key] = pickle.loads(result[key])
